[ANI-68-ani-path-detail] refactor: fsd 아키텍처에 맞게 폴더, 컴포넌트를 분리합니다.
