+++ apps/web/app/path/[pathId]/detail/loading.tsx
@@ -1,3 +1,3 @@
-export default function Loading() {
+export default async function Loading() {
   return <div>패스 상세 데이터 로딩 중...</div>
 }
