+++ apps/web/app/path/[pathId]/load/page.tsx
@@ -1,19 +1,26 @@
-import { getPathLoad } from '@/entities/path/lib/apis'
-import { END_POINT } from '@/shared/config/constants/end-point'
+import { getPathLoad } from '@/views/path/api/get-path-load'
 import { notFound, redirect } from 'next/navigation'
+import { Suspense } from 'react'
 
 type Props = {
   params: { pathId: string }
 }
 
-export default async function Page({ params }: Props) {
-  const data = await getPathLoad(params.pathId)
-
+async function PathLoader({ pathId }: { pathId: string }) {
+  const data = await getPathLoad({ pathId })
   const { status, data: pathData } = data
 
   if (status === 404) {
     return notFound()
   }
 
-  return redirect(END_POINT.PATH.PATH_PAGE(pathData.pathId, pathData.partId, pathData.pageId))
+  redirect(`/path/${pathId}/detail?partId=${pathData.partId}&pageId=${pathData.pageId}`)
+}
+
+export default async function Page({ params: { pathId } }: Props) {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <PathLoader pathId={pathId} />
+    </Suspense>
+  )
 }
