--- apps/web/app/path/[pathId]/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return <div>패스 소개 데이터 로딩 중...</div>
-}
