--- apps/web/src/entities/path/lib/apis/index.ts
@@ -1,4 +1,0 @@
-export { getPathIntroduce } from './get-path-introduce'
-export { getPathPage } from './get-path-page'
-export { getPathSidebarStatus } from './get-path-sidebar-status'
-export { getPathLoad } from './get-path-load'
