--- apps/web/src/entities/path/model/path-detail.ts
@@ -1,13 +1,0 @@
-import { PartType, PageType } from '@/shared/types/common'
-
-type StatusPageType = PageType & {
-  isSuccess: boolean
-}
-
-type StatusPartType = PartType & {
-  page: StatusPageType[]
-}
-
-export type PathDetail = {
-  parts: StatusPartType[]
-}
