--- apps/web/src/entities/path/ui/path-quiz/index.tsx
@@ -1,9 +1,0 @@
-import { GetPathPageResponseType } from '@/entities/path/lib/types/apis'
-
-type Props = {
-  data: GetPathPageResponseType['data'] & { type: 'QUIZ' }
-}
-
-export default function PathQuiz({ data }: Props) {
-  return <div>{JSON.stringify(data.quiz.quizs)}</div>
-}
