+++ apps/web/src/shared/ui/logo/Logo.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import Logo from '.'
 
 const meta = {
-  title: 'Widget/Logo',
+  title: 'Shared/Logo',
   component: Logo,
   parameters: {
     layout: 'centered'
