+++ apps/web/src/views/path/api/index.ts
@@ -1,0 +1,4 @@
+export { getPathIntroduce } from './get-path-introduce'
+export { getPathPage } from './get-path-page'
+export { getPathStatus } from './get-path-status'
+export { getPathLoad } from './get-path-load'
